refactor: 글작성 가독성 리팩토링
