+++ src/main/java/teamexpress/velo9/post/controller/PostController.java
@@ -18,12 +18,11 @@ import teamexpress.velo9.common.controller.BaseController;
 import teamexpress.velo9.common.domain.PostResult;
 import teamexpress.velo9.common.domain.Result;
 import teamexpress.velo9.member.security.oauth.SessionConst;
-import teamexpress.velo9.post.domain.Post;
 import teamexpress.velo9.post.dto.LookPostDTO;
 import teamexpress.velo9.post.dto.LoveDTO;
 import teamexpress.velo9.post.dto.LovePostDTO;
+import teamexpress.velo9.post.dto.PostLoadDTO;
 import teamexpress.velo9.post.dto.PostReadDTO;
-import teamexpress.velo9.post.dto.PostSaveDTO;
 import teamexpress.velo9.post.dto.PostWriteDTO;
 import teamexpress.velo9.post.dto.ReadDTO;
 import teamexpress.velo9.post.dto.SeriesDTO;
@@ -46,16 +45,15 @@ public class PostController extends BaseController {
 	private final TagService tagService;
 
 	@GetMapping("/write")
-	public PostWriteDTO write(@RequestParam(required = false) Long id) {
-		return id == null ? null : postService.findPostById(id);
+	public PostLoadDTO write(@RequestParam(required = false) Long id) {
+		return id != null ? postService.findPostById(id) : null;
 	}
 
 	@PostMapping("/write")
-	public Result write(@Valid @RequestBody PostSaveDTO postSaveDTO, HttpSession session) {
-		Post post = postService.write(postSaveDTO, getMemberId(session));
-		tagService.addTags(post, postSaveDTO.getTags());
+	public Result<Long> write(@Valid @RequestBody PostWriteDTO postWriteDTO, HttpSession session) {
+		Long postId = postService.write(postWriteDTO, getMemberId(session));
 		tagService.removeUselessTags();
-		return new Result<>(post.getId());
+		return new Result<>(postId);
 	}
 
 	@PostMapping("/writeTemporary")
