+++ src/main/java/teamexpress/velo9/post/service/PostService.java
@@ -21,17 +21,18 @@ import teamexpress.velo9.post.domain.PostStatus;
 import teamexpress.velo9.post.domain.PostTag;
 import teamexpress.velo9.post.domain.PostTagRepository;
 import teamexpress.velo9.post.domain.PostThumbnail;
-import teamexpress.velo9.post.domain.PostThumbnailRepository;
 import teamexpress.velo9.post.domain.Series;
 import teamexpress.velo9.post.domain.SeriesRepository;
+import teamexpress.velo9.post.domain.Tag;
+import teamexpress.velo9.post.domain.TagRepository;
 import teamexpress.velo9.post.domain.TemporaryPost;
 import teamexpress.velo9.post.domain.TemporaryPostRepository;
 import teamexpress.velo9.post.dto.LookPostDTO;
 import teamexpress.velo9.post.dto.LoveDTO;
 import teamexpress.velo9.post.dto.LovePostDTO;
+import teamexpress.velo9.post.dto.PostLoadDTO;
 import teamexpress.velo9.post.dto.PostMainDTO;
 import teamexpress.velo9.post.dto.PostReadDTO;
-import teamexpress.velo9.post.dto.PostSaveDTO;
 import teamexpress.velo9.post.dto.PostWriteDTO;
 import teamexpress.velo9.post.dto.ReadDTO;
 import teamexpress.velo9.post.dto.SearchCondition;
@@ -54,37 +55,108 @@ public class PostService {
 	private final LoveRepository loveRepository;
 	private final LookRepository lookRepository;
 	private final TemporaryPostRepository temporaryPostRepository;
-	private final PostThumbnailRepository postThumbnailRepository;
+	private final TagRepository tagRepository;
 	private final PostTagRepository postTagRepository;
 
 	@Transactional
-	public Post write(PostSaveDTO postSaveDTO, Long memberId) {
-		PostThumbnail postThumbnail = getPostThumbnail(postSaveDTO.getThumbnailFileName());
-		Series series = getSeries(postSaveDTO.getSeriesId());
+	public Long write(PostWriteDTO postWriteDTO, Long memberId) {
+		if (isNewPost(postWriteDTO)) {
+			return writeNew(postWriteDTO, memberId);
+		}
+		return edit(postWriteDTO, memberId);
+	}
+
+	private boolean isNewPost(PostWriteDTO postWriteDTO) {
+		return postWriteDTO.getPostId() == null;
+	}
+
+	private Long writeNew(PostWriteDTO postWriteDTO, Long memberId) {
+		PostThumbnail postThumbnail = getPostThumbnail(postWriteDTO.getThumbnailFileName());
+		Series series = getSeries(postWriteDTO.getSeriesId());
 		Member member = getMember(memberId);
-		if (postThumbnail != null) {
-			postThumbnailRepository.save(postThumbnail);
+
+		Post post = postRepository.save(postWriteDTO.toPost(member, series, postThumbnail));
+
+		updateTags(post, postWriteDTO.getTags());
+
+		return post.getId();
+	}
+
+	private Long edit(PostWriteDTO postWriteDTO, Long memberId) {
+		PostThumbnail postThumbnail = getPostThumbnail(postWriteDTO.getThumbnailFileName());
+		Series series = getSeries(postWriteDTO.getSeriesId());
+
+		Post post = postRepository.findById(postWriteDTO.getPostId()).orElseThrow();
+
+		checkSameWriter(post, memberId);
+
+		postWriteDTO.makeIntroduce();
+
+		post.edit(postWriteDTO.getTitle(),
+			postWriteDTO.getIntroduce(),
+			postWriteDTO.getContent(),
+			postWriteDTO.getAccess(),
+			series,
+			postThumbnail);
+
+		updateTags(post, postWriteDTO.getTags());
+
+		return post.getId();
+	}
+
+	private PostThumbnail getPostThumbnail(String fileName) {
+		return fileName != null ? PostThumbnailFileUploader.divideFileName(fileName).toPostThumbnail() : null;
+	}
+
+	private Series getSeries(Long seriesId) {
+		return seriesId != null ? seriesRepository.findById(seriesId).orElse(null) : null;
+	}
+
+	private Member getMember(Long memberId) {
+		if (memberId == null) {
+			throw new NullPointerException("member id MUST NOT BE NULL!!!");
 		}
 
-		Post post = null;
+		return memberRepository.findById(memberId)
+			.orElseThrow(() -> new NullPointerException("no member"));
+	}
 
-		if (postSaveDTO.getPostId() == null) {
-			post = postRepository.save(postSaveDTO.toPost(member, series, postThumbnail));
+	private void updateTags(Post post, List<String> tags) {
+		postTagRepository.deleteAllByPost(post);
+
+		if (isEmpty(tags)) {
+			return;
 		}
 
-		if (postSaveDTO.getPostId() != null) {
-			post = postRepository.findById(postSaveDTO.getPostId()).orElseThrow();
-			checkSameMember(post,memberId);
-			postSaveDTO.setIntroduce();
-			post.edit(postSaveDTO.getTitle(),
-				postSaveDTO.getIntroduce(),
-				postSaveDTO.getContent(),
-				postSaveDTO.getAccess(),
-				series,
-				postThumbnail);
-		}
+		tags = removeDuplication(tags);
 
-		return post;
+		List<String> realTags = tagRepository.getTagNames();
+
+		tags.stream().filter(name -> !realTags.contains(name))
+			.forEach(name -> tagRepository.save(
+				Tag.builder().name(name).build()
+			));
+
+		tags.forEach(name -> postTagRepository.save(
+			PostTag.builder()
+				.tag(tagRepository.findByName(name))
+				.post(post)
+				.build()
+		));
+	}
+
+	private void checkSameWriter(Post post, Long memberId) {
+		if (!post.getMember().getId().equals(memberId)) {
+			throw new IllegalStateException("잘못된 접근입니다.");
+		}
+	}
+
+	private List<String> removeDuplication(List<String> tags) {
+		return tags.stream().distinct().collect(Collectors.toList());
+	}
+
+	private boolean isEmpty(List<String> tags) {
+		return tags == null || tags.isEmpty();
 	}
 
 	@Transactional
@@ -103,11 +175,11 @@ public class PostService {
 		postRepository.deleteById(id);
 	}
 
-	public PostWriteDTO findPostById(Long id) {
+	public PostLoadDTO findPostById(Long id) {
 		Post post = postRepository.findById(id).orElseThrow(
 			() -> new IllegalStateException("존재하지 않는 포스트입니다."));
 		List<PostTag> postTags = postTagRepository.findByPost(post);
-		return new PostWriteDTO(post, postTags);
+		return new PostLoadDTO(post, postTags);
 	}
 
 	public Slice<SeriesDTO> findSeries(String nickname, Pageable pageable) {
@@ -145,27 +217,6 @@ public class PostService {
 		return findPosts.stream()
 			.map(TempSavedPostDTO::new)
 			.collect(Collectors.toList());
-	}
-
-	private PostThumbnail getPostThumbnail(String fileName) {
-		PostThumbnail postThumbnail = null;
-		if (fileName != null) {
-			postThumbnail = PostThumbnailFileUploader.divideFileName(fileName).toPostThumbnail();
-		}
-		return postThumbnail;
-	}
-
-	private Member getMember(Long memberId) {
-		if (memberId == null) {
-			throw new NullPointerException("member id MUST NOT BE NULL!!!");
-		}
-
-		return memberRepository.findById(memberId)
-			.orElseThrow(() -> new NullPointerException("no member"));
-	}
-
-	private Series getSeries(Long seriesId) {
-		return seriesId == null ? null : seriesRepository.findById(seriesId).orElse(null);
 	}
 
 	private Long writeAlternativeTemporary(TemporaryPostWriteDTO temporaryPostWriteDTO, Long memberId) {
@@ -249,12 +300,6 @@ public class PostService {
 	private void checkOwner(Post findPost, String nickname) {
 		if (!findPost.getMember().getNickname().equals(nickname)) {
 			throw new IllegalStateException("비정상적인 접근입니다.");
-		}
-	}
-
-	private void checkSameMember(Post post, Long memberId){
-		if(post.getMember().getId() != memberId){
-			throw new IllegalStateException("잘못된 접근입니다.");
 		}
 	}
 }
